# docs/openapi/flowcept-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flowcept Webservice API",
    "description": "Read-only REST API for Flowcept provenance data. Provides workflows, tasks, and objects endpoints with query support.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "List Campaigns",
        "description": "List derived campaign summaries, most recently active first.",
        "operationId": "list_campaigns_api_v1_campaigns_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Get Campaign",
        "description": "Get one campaign: derived summary, its workflows, and a task summary.",
        "operationId": "get_campaign_api_v1_campaigns__campaign_id__get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Campaign Api V1 Campaigns  Campaign Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/provenance_card": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "Get Campaign Provenance Card",
        "description": "Get a campaign provenance card as structured JSON or rendered markdown.",
        "operationId": "get_campaign_provenance_card_api_v1_campaigns__campaign_id__provenance_card_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "description": "List derived agent summaries, most recently active first.",
        "operationId": "list_agents_api_v1_agents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent",
        "description": "Get one agent's derived summary and per-activity task summary.",
        "operationId": "get_agent_api_v1_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Agent Api V1 Agents  Agent Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{agent_id}/tasks": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Tasks",
        "description": "List tasks executed by or sent from an agent.",
        "operationId": "get_agent_tasks_api_v1_agents__agent_id__tasks_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboards": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "List Dashboards",
        "description": "List all stored dashboards.",
        "operationId": "list_dashboards_api_v1_dashboards_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dashboards"
        ],
        "summary": "Create Dashboard",
        "description": "Create a dashboard; the server assigns its id and timestamps.",
        "operationId": "create_dashboard_api_v1_dashboards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Dashboard Api V1 Dashboards Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/{dashboard_id}": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "Get Dashboard",
        "description": "Get a dashboard by id.",
        "operationId": "get_dashboard_api_v1_dashboards__dashboard_id__get",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dashboard Api V1 Dashboards  Dashboard Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "dashboards"
        ],
        "summary": "Update Dashboard",
        "description": "Replace a dashboard spec, preserving its id and creation time.",
        "operationId": "update_dashboard_api_v1_dashboards__dashboard_id__put",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardSpec"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Dashboard Api V1 Dashboards  Dashboard Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "dashboards"
        ],
        "summary": "Delete Dashboard",
        "description": "Delete a dashboard by id.",
        "operationId": "delete_dashboard_api_v1_dashboards__dashboard_id__delete",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Dashboard Api V1 Dashboards  Dashboard Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Card": {
        "properties": {
          "card_id": {
            "type": "string",
            "title": "Card Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "chart",
              "metric",
              "table",
              "markdown",
              "prov_card"
            ],
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "live": {
            "type": "boolean",
            "title": "Live",
            "default": false
          },
          "refresh_interval_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Interval Sec"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardData"
              },
              {
                "type": "null"
              }
            ]
          },
          "viz": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VizSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "type"
        ],
        "title": "Card",
        "description": "One dashboard card. Content fields depend on ``type``."
      },
      "CardData": {
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "tasks",
              "workflows",
              "objects"
            ],
            "title": "Source",
            "default": "tasks"
          },
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "group_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group By"
          },
          "metrics": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MetricSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "x": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "sort": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort"
          },
          "limit": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 500
          }
        },
        "type": "object",
        "title": "CardData",
        "description": "Declarative data binding for a card: what to query and how to shape it."
      },
...
      "DashboardSpec": {
        "properties": {
          "dashboard_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashboard Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/Card"
            },
            "type": "array",
            "title": "Cards"
          },
          "layout": {
            "items": {
              "$ref": "#/components/schemas/LayoutItem"
            },
            "type": "array",
            "title": "Layout"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DashboardSpec",
        "description": "A complete dashboard: context filter, cards, and layout."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LayoutItem": {
        "properties": {
          "card_id": {
            "type": "string",
            "title": "Card Id"
          },
          "x": {
            "type": "integer",
            "maximum": 11.0,
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y"
          },
          "w": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "W"
          },
          "h": {
            "type": "integer",
            "minimum": 1.0,
            "title": "H"
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "LayoutItem",
        "description": "Grid placement of a card in a 12-column layout."
      },
      "ListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count",
          "limit"
        ],
        "title": "ListResponse",
        "description": "Generic list envelope for collection endpoints."
      },
      "MetricSpec": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "agg": {
            "type": "string",
            "enum": [
              "avg",
              "sum",
              "min",
              "max",
              "count"
            ],
            "title": "Agg"
          }
        },
        "type": "object",
        "required": [
          "field",
          "agg"
        ],
        "title": "MetricSpec",
        "description": "A single aggregation over a (dot-notated) field."
      },
...
      "SortSpec": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          },
          "order": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Order",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortSpec",
        "description": "Sort field/order pair."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VizSpec": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "line",
              "bar",
              "pie",
              "scatter",
              "area",
              "heatmap"
            ],
            "title": "Kind",
            "default": "line"
          },
          "stacked": {
            "type": "boolean",
            "title": "Stacked",
            "default": false
          }
        },
        "type": "object",
        "title": "VizSpec",
        "description": "How a chart card renders its rows."
      }
    }
  }
}